adjust seeder, making components, and some post item detail

+++ database/migrations/2024_10_07_133221_create_properties_table.php
@@ -30,7 +30,7 @@ return new class extends Migration
             $table->smallInteger('property_direction')->nullable();   // huong nha
             $table->smallInteger('property_legal')->nullable();         // phap ly
             $table->smallInteger('property_status')->nullable();        // tinh trang
-            $table->mediumInteger('property_price')->default(0);        // gia tien - gia tri cua mediumInt https://dev.to/kakisoft/laravel-mysql-columntype-is-set-to-int-11-even-though-the-size-of-int-was-specified-59pj
+            $table->integer('property_price')->default(0);        // gia tien - gia tri cua mediumInt https://dev.to/kakisoft/laravel-mysql-columntype-is-set-to-int-11-even-though-the-size-of-int-was-specified-59pj
 
             // BAN DO
             $table->string('property_latitude')->nullable();
